refactor: 폴더명 수정
- 줌 기반에서 델타 영역으로 수정되면서 이름을 바꿈

[#831]
