--- frontend/src/components/google-maps/marker/HighZoomMarkerContainer/index.ts
@@ -1,3 +1,0 @@
-import HighZoomMarkerContainer from './HighZoomMarkerContainer';
-
-export default HighZoomMarkerContainer;
