+++ frontend/src/components/google-maps/marker/LargeDeltaAreaMarkerContainer/index.ts
@@ -1,0 +1,3 @@
+import LargeDeltaAreaMarkerContainer from './LargeDeltaAreaMarkerContainer';
+
+export default LargeDeltaAreaMarkerContainer;
