--- frontend/src/components/google-maps/marker/LowZoomMarkerContainer/index.ts
@@ -1,3 +1,0 @@
-import LowZoomMarkerContainer from './LowZoomMarkerContainer';
-
-export default LowZoomMarkerContainer;
