+++ frontend/src/components/google-maps/marker/MaxDeltaAreaMarkerContainer/MaxDeltaAreaMarkerContainer.tsx
@@ -1,7 +1,7 @@
 import RegionMarkerRenderer from './components/RegionMarkerRenderer';
 import { useRegionMarkers } from './hooks/useRegionMarkers';
 
-const LowZoomMarkerContainer = () => {
+const MaxDeltaAreaMarkerContainer = () => {
   const { data: regions, isSuccess, isError } = useRegionMarkers();
   if (!regions || !isSuccess || isError) {
     return <></>;
@@ -10,4 +10,4 @@ const LowZoomMarkerContainer = () => {
   return regions.map((region) => <RegionMarkerRenderer key={region.regionName} region={region} />);
 };
 
-export default LowZoomMarkerContainer;
+export default MaxDeltaAreaMarkerContainer;
