+++ frontend/src/components/google-maps/marker/MaxDeltaAreaMarkerContainer/components/RegionMarker.stories.tsx
