+++ frontend/src/components/google-maps/marker/MaxDeltaAreaMarkerContainer/components/RegionMarkerRenderer.tsx
