+++ frontend/src/components/google-maps/marker/MaxDeltaAreaMarkerContainer/index.ts
@@ -1,0 +1,3 @@
+import MaxDeltaAreaMarkerContainer from './MaxDeltaAreaMarkerContainer';
+
+export default MaxDeltaAreaMarkerContainer;
