--- frontend/src/components/google-maps/marker/MiddleZoomMarkerContainer/index.ts
@@ -1,3 +1,0 @@
-import MiddleZoomMarkerContainer from './MiddleZoomMarkerContainer';
-
-export default MiddleZoomMarkerContainer;
