+++ frontend/src/components/google-maps/marker/SmallMediumDeltaAreaMarkerContainer/constants/index.ts
