+++ frontend/src/components/google-maps/marker/SmallMediumDeltaAreaMarkerContainer/index.ts
@@ -1,0 +1,3 @@
+import SmallMediumDeltaAreaMarkerContainer from './SmallMediumDeltaAreaMarkerContainer';
+
+export default SmallMediumDeltaAreaMarkerContainer;
