+++ frontend/src/stores/google-maps/deltaAreaStore/deltaAreaStore.test.ts
