+++ frontend/src/stores/google-maps/deltaAreaStore/index.ts
